Update azure-to-azure-how-to-reprotect.md
Miss-leading, changed the wording at the top: "If you fail back the VMs to the primary region, you need to do the following" to "In order to fail back the VMs to the primary region, you need to do the following". As previously is sounds like you have already done the fail back.